student-sql/Неделя5/requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
apache-airflow==3.2.1
apache-airflow-core==3.2.1
apache-airflow-providers-amazon==9.27.0
apache-airflow-providers-common-compat==1.14.3
apache-airflow-providers-common-io==1.7.2
apache-airflow-providers-common-sql==1.35.0
apache-airflow-providers-http==6.0.2
apache-airflow-providers-postgres==6.6.3
apache-airflow-providers-smtp==3.0.0
apache-airflow-providers-standard==1.12.3
apache-airflow-task-sdk==1.2.1
argcomplete==3.6.3
asgiref==3.11.1
asn1crypto==1.5.1
asyncpg==0.31.0
attrs==26.1.0
awscli-local==0.22.2
babel==2.18.0
beautifulsoup4==4.14.3
boto3==1.43.2
boto3-stubs==1.43.2
botocore==1.43.2
botocore-stubs==1.42.41
cachetools==7.1.0
cadwyn==6.2.2
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorlog==6.10.1
cron_descriptor==2.0.8
croniter==6.2.2
cryptography==47.0.0
Deprecated==1.3.1
dill==0.4.1
dnslib==0.9.26
dnspython==2.8.0
docker==7.1.0
email-validator==2.3.0
fastapi==0.136.1
fastapi-cli==0.0.24
frozenlist==1.8.0
fsspec==2026.4.0
googleapis-common-protos==1.74.0
graphviz==0.21
greenback==1.3.0
greenlet==3.5.0
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
importlib_metadata==8.7.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
jsonpath-ng==1.8.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.1.0
localstack==2026.4.0
localstack-client==2.11
lockfile==0.12.2
loguru==0.7.3
lxml==6.0.2
Mako==1.3.12
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.3.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==11.0.2
msgspec==0.21.1
multidict==6.7.1
natsort==8.4.0
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-exporter-otlp-proto-http==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
outcome==1.3.0.post0
packaging==26.2
pathlib_abc==0.5.2
pathspec==1.1.1
pendulum==3.2.0
pluggy==1.6.0
plux==1.16.0
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.12
PyAthena==3.30.1
pycparser==3.0
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
pygtrie==2.5.0
PyJWT==2.12.1
pyotp==2.9.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
python-slugify==8.0.4
python-telegram-bot==22.7
pytz==2026.1.post1
PyYAML==6.0.3
redshift_connector==2.1.13
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rich-argparse==1.8.0
rich-toolkit==0.19.7
rpds-py==0.30.0
s3transfer==0.17.0
sagemaker_studio==1.0.26
scramp==1.4.8
semver==3.0.4
setproctitle==1.3.7
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==0.52.1
structlog==25.5.0
svcs==25.1.0
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
typer==0.25.1
types-awscrt==0.31.3
types-s3transfer==0.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
uc-micro-py==2.0.0
universal_pathlib==0.3.10
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
watchtower==3.4.0
websockets==16.0
wirerope==1.0.0
wrapt==2.1.2
yarl==1.23.0
zipp==3.23.1
